+++ CHANGELOG.md
@@ -1,8 +1,9 @@
+![Advancements](markdown_assets/advancements.png)
+
 ## 3.2: Advancements update
 
 ## New features
-- Added new advancements:
-- 
+- Added 9 new advancements
 
 ## Bug fixes
 - Crash when reloading assets related to shaders