+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 3.4
 
+## New features
+- Better health rendering. This improves compatibility with resource packs which changes heart textures. (bleudev [#67](https://github.com/bleudev/nine_lifes/pull/67))
+
 ## Optimisations
-- Reduced jar file size by compressing png files
+- Reduced jar file size by compressing png files (bleudev [#67](https://github.com/bleudev/nine_lifes/pull/67))