+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 3.4
 
+## Breaking changes
+- `ServerPlayer.setLifes(lifesCountChanger)` was deleted as deprecated in 3.2. Use `ServerPlayer.lifes` property instead.
+- `ServerPlayer.addLifes(addedLifesCount)` was deleted as deprecated in 3.2. Use `ServerPlayer.lifes` property incrementing instead.
+
 ## New features
 - Better health rendering. This improves compatibility with resource packs which changes heart textures. (bleudev [#67](https://github.com/bleudev/nine_lifes/pull/67))
 - Insomnia Potions (bleudev [#68](https://github.com/bleudev/nine_lifes/pull/68))